refactor: useReport fetchurl 변경
