+++ src/app/tanstack-query/reports/useReports.ts
@@ -1,15 +1,15 @@
-import {SESSION_STORAGE_KEY_TOKEN} from "@api/keys";
-import {DOMAIN} from "@api/url";
-import {getSessionStorage} from "@app/utils/storage";
-import {QUERY_KEY_REPORT} from "@constants/queryKeys";
-import {useQuery} from "@tanstack/react-query";
-import {MonthScheduleQuery,} from "@app/types/schedule.ts";
-import {Report} from "@app/types/report.ts";
+import { SESSION_STORAGE_KEY_TOKEN } from "@api/keys.ts";
+import { DOMAIN } from "@api/url.ts";
+import { getSessionStorage } from "@utils/storage.ts";
+import { QUERY_KEY_REPORT } from "@constants/queryKeys.ts";
+import { useQuery } from "@tanstack/react-query";
+import { MonthScheduleQuery } from "@app/types/schedule.ts";
+import { Report } from "@app/types/report.ts";
 
 const fetchReport = async (query: MonthScheduleQuery) => {
   const token = getSessionStorage(SESSION_STORAGE_KEY_TOKEN, "");
 
-  return fetch(`${DOMAIN}/report`, {
+  return fetch(`${DOMAIN}/report/home`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
